+++ Content.Shared/Actions/SharedActionsSystem.cs
@@ -657,7 +657,7 @@ public abstract class SharedActionsSystem : EntitySystem
         var worldValidated
             = ValidateWorldTargetBase(user, coords, comp.CheckCanInteract, comp.CheckCanAccess, comp.Range);
 
-        if (!entityValidated || !worldValidated) //CP14 OR not AND (experiment)
+        if (!entityValidated && !worldValidated)
             return false;
 
         var ev = new ValidateActionEntityWorldTargetEvent(user,
