+++ Resources/Prototypes/_CP14/Entities/Actions/Spells/Meta/T0_mana_gift.yml
@@ -26,7 +26,7 @@
   - type: CP14MagicEffectSomaticAspect
   - type: CP14MagicEffectCastingVisual
     proto: CP14RuneManaGift
-  - type: EntityWorldTargetAction
+  - type: EntityTargetAction
     whitelist:
       components:
       - CP14MagicEnergyContainer
@@ -38,7 +38,7 @@
     icon:
       sprite: _CP14/Effects/Magic/spells_icons.rsi
       state: mana_gift
-    event: !type:CP14DelayedEntityWorldTargetActionEvent
+    event: !type:CP14DelayedEntityTargetActionEvent
       cooldown: 5
       castDelay: 2
