+++ Resources/Prototypes/_CP14/Entities/Actions/Spells/Movement/T1_shadow_grab.yml
@@ -20,7 +20,7 @@
       spawns:
       - CP14ImpactEffectShadowStep
   - type: CP14MagicEffectSomaticAspect
-  - type: EntityWorldTargetAction
+  - type: EntityTargetAction
     whitelist:
       components:
       - Item
@@ -33,7 +33,7 @@
     icon:
       sprite: _CP14/Effects/Magic/spells_icons.rsi
       state: shadow_grab
-    event: !type:CP14DelayedEntityWorldTargetActionEvent
+    event: !type:CP14DelayedEntityTargetActionEvent
       cooldown: 10
       castDelay: 0
       entityDistance: 1000
